Split out unsafe and eval parts of top-level
The top-level module was a mix of code for munging and outputting
information about migrations, and code that actually executed
migrations. The former has been split out into
Database.Beam.AutoMigrate.Eval, and the latter into
Database.Beam.Automigrate.Unsafe.

The top-level module now exports migration executions that wrap
everything in a transaction, and the Unsafe module exports them without
the transaction wrapper: that's why they're extremely unsafe.
